talk-to-your-resume-bot/requirements.txt
langchain
openai
langchain-community
langchain-openrouter
python-dotenv
streamlit
pydantic
pypdf
